+++ src/modules/sync/components/sync-connected-sites.tsx
@@ -11,7 +11,6 @@ import { CircleRedCrossIcon } from 'src/components/icons/circle-red-cross';
 import offlineIcon from 'src/components/offline-icon';
 import { PressableLogo } from 'src/components/pressable-logo';
 import ProgressBar from 'src/components/progress-bar';
-import { SyncPullPushClear } from 'src/components/sync-pull-push-clear';
 import { Tooltip, DynamicTooltip } from 'src/components/tooltip';
 import { WordPressLogoCircle } from 'src/components/wordpress-logo-circle';
 import { useSyncSites } from 'src/hooks/sync-sites';
@@ -25,6 +24,7 @@ import { getIpcApi } from 'src/lib/get-ipc-api';
 import { getLocalizedLink } from 'src/lib/get-localized-link';
 import { OpenSitesSyncSelector } from 'src/modules/sync';
 import { SyncDialog } from 'src/modules/sync/components/sync-dialog';
+import { SyncPullPushClear } from 'src/modules/sync/components/sync-pull-push-clear';
 import { convertTreeToOptionsToSync } from 'src/modules/sync/convert-tree-to-options-to-sync';
 import { useI18nLocale } from 'src/stores';
 import type { SyncSite } from 'src/hooks/use-fetch-wpcom-sites/types';
