+++ src/modules/sync/components/sync-tab-image.tsx
