+++ src/modules/sync/index.tsx
@@ -12,7 +12,7 @@ import { getIpcApi } from 'src/lib/get-ipc-api';
 import { ConnectButton } from 'src/modules/sync/components/connect-button';
 import { SyncConnectedSites } from 'src/modules/sync/components/sync-connected-sites';
 import { SyncSitesModalSelector } from 'src/modules/sync/components/sync-sites-modal-selector';
-import { SyncTabImage } from 'src/modules/sync/images/sync-tab-image';
+import { SyncTabImage } from 'src/modules/sync/components/sync-tab-image';
 import type { SyncSite } from 'src/hooks/use-fetch-wpcom-sites/types';
 
 function SiteSyncDescription( { children }: PropsWithChildren ) {
